+++ src/api/index.js
@@ -1,0 +1,3 @@
+export { default as request } from './request'
+export { TOKEN_KEY } from '../store/auth'
+export * from './user'